Move buildWhichOrderDynamicList to MidEventHandler
