+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/judgeapproval/ApproveOrderService.java
@@ -1,28 +1,16 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.service.judgeapproval;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.tika.utils.StringUtils;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DynamicList;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.DynamicListElement;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.judgeapproval.JudgeApproval;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.wrapper.DraftOrdersWrapper;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import static java.lang.String.format;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.judgeapproval.JudgeApprovalDocType.DRAFT_ORDER;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.judgeapproval.JudgeApprovalDocType.PSA;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.judgeapproval.JudgeDecision.JUDGE_NEEDS_TO_MAKE_CHANGES;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.judgeapproval.JudgeDecision.READY_TO_BE_SEALED;
 
 @Service
 @Slf4j
 public class ApproveOrderService {
-
-    protected static final String SEPARATOR = "#";
 
     private final JudgeApprovalResolver judgeApprovalResolver;
 
@@ -61,61 +49,6 @@ public class ApproveOrderService {
         return doc;
     }
 
-    /**
-     * Builds a {@link DynamicList} representing the orders available for selection
-     * based on the given {@link DraftOrdersWrapper}. Each order is represented as a
-     * {@link DynamicListElement} containing a code and label.
-     *
-     * <p>The method processes up to 5 judge approvals, where each approval contributes a list item
-     * if the approval is non-null and the associated document filename is not empty.</p>
-     *
-     * <p>For each {@link JudgeApproval}:</p>
-     * <ul>
-     *     <li>The code is prefixed with "draftOrder" or "psa", depending on the document type.</li>
-     *     <li>The filename is used as the label of the list item.</li>
-     * </ul>
-     *
-     * <p>Example:</p>
-     * <pre>
-     * Code: draftOrder_1, Label: OrderDocument1.pdf
-     * Code: psa_2, Label: PensionSharingAnnex1.pdf
-     * </pre>
-     *
-     * @param draftOrdersWrapper the {@link DraftOrdersWrapper} containing the judge approvals.
-     * @return a {@link DynamicList} populated with items representing the orders.
-     */
-    public DynamicList buildWhichOrderDynamicList(DraftOrdersWrapper draftOrdersWrapper) {
-        List<DynamicListElement> listItems = new ArrayList<>();
-
-        for (int i = 1; i <= 5; i++) {
-            JudgeApproval judgeApproval = resolveJudgeApproval(draftOrdersWrapper, i);
-            if (judgeApproval != null) {
-                String codePrefix = DRAFT_ORDER == judgeApproval.getDocType() ? DRAFT_ORDER.name() : PSA.name();
-                String code = codePrefix + SEPARATOR + i;
-
-                String filename = getDocumentFileName(judgeApproval);
-                if (!StringUtils.isEmpty(filename)) {
-                    listItems.add(DynamicListElement.builder().code(code).label(filename).build());
-                }
-            }
-        }
-
-        return DynamicList.builder().listItems(listItems).build();
-    }
-
-    private String getDocumentFileName(JudgeApproval judgeApproval) {
-        String filename = null;
-        if (JUDGE_NEEDS_TO_MAKE_CHANGES == judgeApproval.getJudgeDecision()) {
-            filename = judgeApproval.getAmendedDocument() != null
-                ? judgeApproval.getAmendedDocument().getDocumentFilename()
-                : "Unknown Filename";
-        } else if (READY_TO_BE_SEALED == judgeApproval.getJudgeDecision()) {
-            filename = judgeApproval.getDocument() != null
-                ? judgeApproval.getDocument().getDocumentFilename()
-                : "Unknown Filename";
-        }
-        return filename;
-    }
 
     /**
      * Resolves the {@link JudgeApproval} object from the provided {@link DraftOrdersWrapper}
@@ -145,5 +78,4 @@ public class ApproveOrderService {
             default -> null;
         };
     }
-
 }
